chore(endpoints): refactor materialization module
